small components added
