+++ src/components/data/InfoData.tsx
@@ -1,0 +1,22 @@
+import wallet from "../../assets/howItWorks/wallet.png";
+import collection from "../../assets/howItWorks/collection.png";
+import earning from "../../assets/howItWorks/earning.png";
+
+export const infoCardsData = [
+    {
+      image: wallet,
+      title: "Setup Your Wallet",
+      description: "Set up your wallet of choice. Connect it to the NFT market by clicking the wallet icon in the top right corner."
+    },
+    {
+      image: collection,
+      title: "Create Collection",
+      description: "Upload your work and setup your collection. Add a description, social links and floor price."
+    },
+    {
+      image: earning,
+      title: "Start Earning",
+      description: "Choose between auctions and fixed-price listings. Start earning by selling your NFTs or trading others."
+    }
+  ];
+
